backend: update user refactor

+++ backend/src/modules/users/services/user.service.ts
@@ -1,7 +1,5 @@
 import { events } from '@events';
-import { errorKeys } from '@exceptions';
-import { TranslatableHttpException } from '@exceptions/TranslatableHttpException';
-import { ResetPasswordTokensRepository } from '@modules/auth/repositories/reset-password-tokens.repository';
+import { BadRequestException, errorKeys } from '@exceptions';
 import { BaseService, userToIUser } from '@modules/common';
 import {
   ActivateUserReqDto,
@@ -9,8 +7,12 @@ import {
   CreateUserReqDto,
   DeactivateUserReqDto,
   DeleteUserReqDto,
+  GetUserReqDto,
+  IGetUserDto,
   IUserDto,
   UnlockUserReqDto,
+  UpdateUserDto,
+  UpdateUserReqDto,
   UserActivatedEvent,
   UserCreatedEvent,
   UserDeactivatedEvent,
@@ -18,30 +20,26 @@ import {
   UserRepository,
   UserRetrievedEvent,
   UserUnlockedEvent,
+  UserUpdatedEvent,
 } from '@modules/users';
-import { isNullOrUndefined } from '@utils';
-import StatusCode from 'status-code-enum';
+import { isNullOrEmptyString, isNullOrUndefined } from '@utils';
 import { Container, Service } from 'typedi';
-import { GetUserReqDto } from '../dtos/get-user.dto';
 import { User } from '../entities/user.entity';
-import { IGetUserDto } from '../interfaces/get-user.interfaces';
 
 @Service()
 export class UsersService extends BaseService {
   private readonly _userRepository: UserRepository;
-  private readonly _resetPasswordTokensRepository: ResetPasswordTokensRepository;
 
   public constructor() {
     super();
     this._userRepository = Container.get(UserRepository);
-    this._resetPasswordTokensRepository = Container.get(ResetPasswordTokensRepository);
   }
 
   public async get(reqDto: GetUserReqDto): Promise<IGetUserDto | null> {
     const user = await this._userRepository.getByUuid(reqDto.userGuid);
 
     if (isNullOrUndefined(user)) {
-      throw new TranslatableHttpException(StatusCode.ClientErrorBadRequest, errorKeys.users.User_Does_Not_Exist, { id: reqDto.userGuid });
+      throw new BadRequestException(errorKeys.users.User_Does_Not_Exist, { id: reqDto.userGuid });
     }
 
     const userDto = this.userToIGetUserDto(user!);
@@ -57,7 +55,7 @@ export class UsersService extends BaseService {
     const userExists = await this._userRepository.checkIfExists({ email: userData?.email, phone: userData?.phone });
 
     if (userExists) {
-      throw new TranslatableHttpException(StatusCode.ClientErrorBadRequest, errorKeys.users.User_Already_Exists, {
+      throw new BadRequestException(errorKeys.users.User_Already_Exists, {
         email: userData.email,
         phone: userData.phone,
       });
@@ -70,48 +68,84 @@ export class UsersService extends BaseService {
     return userDto;
   }
 
-  // public async updateUser(userId: number, userData: UpdateUserDto): Promise<IUserDto> {
-  //   const findUser: IUserDto = await this.users.findUnique({ where: { id: userId } });
-  //   if (!findUser) throw new HttpException(409, "User doesn't exist");
-
-  //   const hashedPassword = await hash(userData.password, 10);
-  //   const updateUserData = await this.users.update({ where: { id: userId }, data: { ...userData, password: hashedPassword } });
-  //   return updateUserData;
-  // }
-
-  public async delete(reqDto: DeleteUserReqDto): Promise<string | null> {
+  public async update(reqDto: UpdateUserReqDto): Promise<IUserDto> {
     const user = await this._userRepository.getByUuid(reqDto.userGuid);
 
     if (isNullOrUndefined(user)) {
-      throw new TranslatableHttpException(StatusCode.ClientErrorBadRequest, errorKeys.users.User_Does_Not_Exist, { id: reqDto.userGuid });
+      throw new BadRequestException(errorKeys.users.User_Does_Not_Exist, { id: reqDto.userGuid });
+    }
+
+    const userData: UpdateUserDto = reqDto.userData;
+
+    let newEmail = userData.email;
+    if (newEmail !== undefined && isNullOrEmptyString(newEmail)) {
+      throw new BadRequestException(errorKeys.users.Invalid_Email, {
+        email: newEmail,
+      });
+    }
+    let newPhone = userData.phone;
+    if (newPhone !== undefined && isNullOrEmptyString(newPhone)) {
+      throw new BadRequestException(errorKeys.users.Invalid_Phone, {
+        phone: newPhone,
+      });
+    }
+
+    if (newEmail === undefined) {
+      newEmail = user!.email;
+    }
+    if (newPhone === undefined) {
+      newPhone = user!.phone;
+    }
+    if (newEmail !== user!.email || newPhone !== user!.phone) {
+      const userExists = await this._userRepository.checkIfExists({ email: newEmail, phone: newPhone }, user!.id);
+
+      if (userExists) {
+        throw new BadRequestException(errorKeys.users.User_Already_Exists, {
+          email: newEmail,
+          phone: newPhone,
+        });
+      }
+    }
+
+    const updatedUser = await this._userRepository.update(user!.id, reqDto);
+    const userDto = userToIUser(updatedUser!);
+    this._eventDispatcher.dispatch(events.users.userUpdated, new UserUpdatedEvent(userDto, reqDto.currentUserId));
+
+    return userDto;
+  }
+
+  public async delete(reqDto: DeleteUserReqDto): Promise<boolean> {
+    const user = await this._userRepository.getByUuid(reqDto.userGuid);
+
+    if (isNullOrUndefined(user)) {
+      throw new BadRequestException(errorKeys.users.User_Does_Not_Exist, { id: reqDto.userGuid });
+    }
+
+    if (user!.isDeleted) {
+      return true;
     }
 
     const relatedData: string[] = await this._userRepository.checkIfCanBeDeleted(user!.id);
 
     if (relatedData.length > 0) {
-      throw new TranslatableHttpException(
-        StatusCode.ClientErrorBadRequest,
-        errorKeys.general.Object_Is_Connected_With_Another_And_Can_Not_Be_Deleted,
-        {
-          id: user?.uuid,
-          relatedData,
-        },
-      );
+      throw new BadRequestException(errorKeys.general.Object_Is_Connected_With_Another_And_Can_Not_Be_Deleted, {
+        id: user!.uuid,
+        relatedData,
+      });
     }
 
-    await this._resetPasswordTokensRepository.deleteTokens(user!.id);
-    await this._userRepository.delete(user!, reqDto);
+    const result = await this._userRepository.delete(user!.id, reqDto);
 
     this._eventDispatcher.dispatch(events.users.userDeleted, new UserDeletedEvent(userToIUser(user!), reqDto.currentUserId));
 
-    return user!.uuid;
+    return result;
   }
 
   public async activate(reqDto: ActivateUserReqDto): Promise<boolean> {
     const user = await this._userRepository.getByUuid(reqDto.userGuid);
 
     if (isNullOrUndefined(user)) {
-      throw new TranslatableHttpException(StatusCode.ClientErrorBadRequest, errorKeys.users.User_Does_Not_Exist, { id: reqDto.userGuid });
+      throw new BadRequestException(errorKeys.users.User_Does_Not_Exist, { id: reqDto.userGuid });
     }
 
     if (user!.isActive) {
@@ -129,7 +163,7 @@ export class UsersService extends BaseService {
     const user = await this._userRepository.getByUuid(reqDto.userGuid);
 
     if (isNullOrUndefined(user)) {
-      throw new TranslatableHttpException(StatusCode.ClientErrorBadRequest, errorKeys.users.User_Does_Not_Exist, { id: reqDto.userGuid });
+      throw new BadRequestException(errorKeys.users.User_Does_Not_Exist, { id: reqDto.userGuid });
     }
 
     if (!user!.isActive) {
@@ -147,7 +181,7 @@ export class UsersService extends BaseService {
     const user = await this._userRepository.getByUuid(reqDto.userGuid);
 
     if (isNullOrUndefined(user)) {
-      throw new TranslatableHttpException(StatusCode.ClientErrorBadRequest, errorKeys.users.User_Does_Not_Exist, { id: reqDto.userGuid });
+      throw new BadRequestException(errorKeys.users.User_Does_Not_Exist, { id: reqDto.userGuid });
     }
 
     if (!user!.isLockedOut) {
